+++ packages/ckeditor5-heading/tests/integration.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2018, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md.
  */
 
@@ -17,7 +17,7 @@ import ClassicTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/classictest
 import { getData as getModelData, setData as setModelData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
 
 describe( 'Heading integration', () => {
-	let editor, doc, element;
+	let editor, model, doc, element;
 
 	beforeEach( () => {
 		element = document.createElement( 'div' );
@@ -29,7 +29,8 @@ describe( 'Heading integration', () => {
 			} )
 			.then( newEditor => {
 				editor = newEditor;
-				doc = editor.document;
+				model = editor.model;
+				doc = model.document;
 			} );
 	} );
 
@@ -42,28 +43,34 @@ describe( 'Heading integration', () => {
 	describe( 'with the enter key', () => {
 		it( 'should make the "enter" command insert a default heading block if the selection ended at the end of a heading block', () => {
 			editor.setData( '<h2>foobar</h2>' );
-			doc.selection.setCollapsedAt( doc.getRoot().getChild( 0 ), 'end' );
+
+			model.change( writer => {
+				writer.setSelection( doc.getRoot().getChild( 0 ), 'end' );
+			} );
 
 			editor.execute( 'enter' );
 
-			expect( getModelData( doc ) ).to.equal( '<heading1>foobar</heading1><paragraph>[]</paragraph>' );
+			expect( getModelData( model ) ).to.equal( '<heading1>foobar</heading1><paragraph>[]</paragraph>' );
 		} );
 
 		it( 'should not alter the "enter" command if selection not ended at the end of a heading block', () => {
 			// This test is to fill code coverage.
 			editor.setData( '<h2>foobar</h2>' );
-			doc.selection.setCollapsedAt( doc.getRoot().getChild( 0 ), 3 );
+
+			model.change( writer => {
+				writer.setSelection( doc.getRoot().getChild( 0 ), 3 );
+			} );
 
 			editor.execute( 'enter' );
 
-			expect( getModelData( doc ) ).to.equal( '<heading1>foo</heading1><heading1>[]bar</heading1>' );
+			expect( getModelData( model ) ).to.equal( '<heading1>foo</heading1><heading1>[]bar</heading1>' );
 		} );
 	} );
 
 	describe( 'with the image feature', () => {
 		// https://github.com/ckeditor/ckeditor5-heading/issues/73
 		it( 'should not destroy the image when a selection converted to a heading', () => {
-			setModelData( editor.document,
+			setModelData( model,
 				'<paragraph>fo[o</paragraph>' +
 				'<image src="foo.png">' +
 					'<caption>xxx</caption>' +
@@ -71,9 +78,9 @@ describe( 'Heading integration', () => {
 				'<paragraph>b]ar</paragraph>'
 			);
 
-			editor.execute( 'heading1' );
+			editor.execute( 'heading', { value: 'heading1' } );
 
-			expect( getModelData( doc ) ).to.equal(
+			expect( getModelData( model ) ).to.equal(
 				'<heading1>fo[o</heading1>' +
 				'<image src="foo.png">' +
 					'<caption>xxx</caption>' +
@@ -85,21 +92,60 @@ describe( 'Heading integration', () => {
 
 	describe( 'with the undo feature', () => {
 		it( 'does not create undo steps when applied to an existing heading (collapsed selection)', () => {
-			setModelData( editor.document, '<heading1>foo[]bar</heading1>' );
+			setModelData( model, '<heading1>foo[]bar</heading1>' );
 
-			editor.execute( 'heading1' );
-			expect( getModelData( editor.document ) ).to.equal( '<heading1>foo[]bar</heading1>' );
+			editor.execute( 'heading', { value: 'heading1' } );
+			expect( getModelData( model ) ).to.equal( '<heading1>foo[]bar</heading1>' );
 
 			expect( editor.commands.get( 'undo' ).isEnabled ).to.be.false;
 		} );
 
 		it( 'does not create undo steps when applied to an existing heading (non–collapsed selection)', () => {
-			setModelData( editor.document, '<heading1>[foo</heading1><heading1>bar]</heading1>' );
+			setModelData( model, '<heading1>[foo</heading1><heading1>bar]</heading1>' );
 
-			editor.execute( 'heading1' );
-			expect( getModelData( editor.document ) ).to.equal( '<heading1>[foo</heading1><heading1>bar]</heading1>' );
+			editor.execute( 'heading', { value: 'heading1' } );
+			expect( getModelData( model ) ).to.equal( '<heading1>[foo</heading1><heading1>bar]</heading1>' );
 
 			expect( editor.commands.get( 'undo' ).isEnabled ).to.be.false;
+		} );
+	} );
+
+	// Remember to sync docs/_snippets/features/custom-heading-elements.js and docs/features/headings.md
+	// with this test when changing it.
+	describe( 'fancy heading sample in the docs', () => {
+		it( 'upcasts the <h2> and <h2 class=fancy> elements when configured to do so', () => {
+			const element = document.createElement( 'div' );
+			document.body.appendChild( element );
+
+			return ClassicTestEditor
+				.create( element, {
+					plugins: [ Paragraph, Heading ],
+					heading: {
+						options: [
+							{ model: 'paragraph', title: 'Paragraph', class: 'ck-heading_paragraph' },
+							{ model: 'heading2', view: 'h2', title: 'Heading 2', class: 'ck-heading_heading2' },
+							{
+								model: 'headingFancy',
+								view: {
+									name: 'h2',
+									class: 'fancy'
+								},
+								title: 'Heading 2 (fancy)',
+								class: 'ck-heading_heading2_fancy',
+								priority: 'high'
+							}
+						]
+					}
+				} )
+				.then( editor => {
+					editor.setData( '<h2>Heading 2</h2><h2 class="fancy">Fancy Heading 2</h2>' );
+
+					expect( editor.getData() )
+						.to.equal( '<h2>Heading 2</h2><h2 class="fancy">Fancy Heading 2</h2>' );
+
+					editor.destroy();
+					element.remove();
+				} );
 		} );
 	} );
 } );