+++ packages/ckeditor5-heading/tests/tickets/40.js
@@ -1,9 +1,9 @@
 /**
- * @license Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2018, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md.
  */
 
-import HeadingEngine from '../../src/headingengine';
+import HeadingEditing from '../../src/headingediting';
 import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor';
 import Enter from '@ckeditor/ckeditor5-enter/src/enter';
 import { getData, setData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
@@ -18,32 +18,32 @@ describe( 'Bug ckeditor5-heading#40', () => {
 	it( 'enter at the end of a heading creates a paragraph, when heading was loaded before enter', () => {
 		return VirtualTestEditor
 			.create( {
-				plugins: [ HeadingEngine, Enter ]
+				plugins: [ HeadingEditing, Enter ]
 			} )
 			.then( newEditor => {
 				editor = newEditor;
 
-				setData( editor.document, '<heading1>foo[]</heading1>' );
+				setData( editor.model, '<heading1>foo[]</heading1>' );
 
 				editor.execute( 'enter' );
 
-				expect( getData( editor.document ) ).to.equal( '<heading1>foo</heading1><paragraph>[]</paragraph>' );
+				expect( getData( editor.model ) ).to.equal( '<heading1>foo</heading1><paragraph>[]</paragraph>' );
 			} );
 	} );
 
 	it( 'enter at the end of a heading creates a paragraph, when enter was loaded before heading', () => {
 		return VirtualTestEditor
 			.create( {
-				plugins: [ Enter, HeadingEngine ]
+				plugins: [ Enter, HeadingEditing ]
 			} )
 			.then( newEditor => {
 				editor = newEditor;
 
-				setData( editor.document, '<heading1>foo[]</heading1>' );
+				setData( editor.model, '<heading1>foo[]</heading1>' );
 
 				editor.execute( 'enter' );
 
-				expect( getData( editor.document ) ).to.equal( '<heading1>foo</heading1><paragraph>[]</paragraph>' );
+				expect( getData( editor.model ) ).to.equal( '<heading1>foo</heading1><paragraph>[]</paragraph>' );
 			} );
 	} );
 } );